pyproject.toml
[tool.poetry]
name = "ozon-env"
version = "0.1.0"
homepage = "https://github.com/archetipo/ozon-env"
description = "Ozon Env API"
readme = "README.md"
authors = ["Alessio Gerace <alessio.gerace@gmail.com>"]
license = "MIT"
packages = [{ include = "ozonenv/**/*.py" }]
exclude = ["tests/*"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pydantic = { version = "^1.9.1" }
httpx = { version = "^0.23.0" }
pytz = { version = "2022.1" }
motor = { version = "3.0.0" }
ujson = { version = "5.3.0" }
python-dateutil = { version = "2.8.2" }
json_logic_qubit = { version = "^0.9.1" }
aiofiles = { version = "^0.8.0" }
aiopath = { version = "^0.6.10" }
jinja2 = { version = ">=3.1.2" }
Babel = { version = ">=2.10.3" }

[tool.poetry.extras]
json_logic = ["json_logic_qubit"]


[tool.poetry.dev-dependencies]
pytest-asyncio = ">=0.19.0"
pytest-mock = "*"
pytest-cov = "2.12.0,<4.0.0"
anyio = ">=0.21"
starlette = ">=0.20.4"
flake9 = ">=3.8.3"
isort = ">=5.8.0"
black = ">=20.8b1"
bandit = ">=1.7.0"

[tool.black]
line_length = 79

[tool.flake8]
max-complexity = 15
max-line-length = 79
extend-ignore = "E203"

[tool.isort]
line_length = 79
include_trailing_comma = true
force_single_line = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore:function ham\(\) is deprecated:DeprecationWarning'
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"